refactor(github): share instrumented GitHub HTTP in GitHubIntegrationBase
