+++ posthog/models/integration.py
@@ -4,10 +4,9 @@ import time
 import base64
 import socket
 import hashlib
-from collections.abc import Mapping
 from dataclasses import dataclass
 from datetime import datetime, timedelta
-from typing import TYPE_CHECKING, Any, Literal, NamedTuple, Optional, cast
+from typing import TYPE_CHECKING, Any, Literal, NamedTuple, Optional
 from urllib.parse import urlencode, urlparse
 
 from products.workflows.backend.providers import MAILDEV_MOCK_DNS_RECORDS
@@ -16,19 +15,17 @@ if TYPE_CHECKING:
     import aiohttp
 
 from django.conf import settings
-from django.core.cache import cache
 from django.db import models
 from django.http import HttpRequest
 from django.utils import timezone
 
-import jwt
 import requests
 import structlog
 from disposable_email_domains import blocklist as disposable_email_domains_list
 from free_email_domains import whitelist as free_email_domains_list
 from google.auth.transport.requests import Request as GoogleRequest
 from google.oauth2 import service_account
-from prometheus_client import Counter, Gauge
+from prometheus_client import Counter
 from requests.auth import HTTPBasicAuth
 from rest_framework import status
 from rest_framework.exceptions import ValidationError
@@ -41,6 +38,7 @@ from stripe import StripeClient
 from posthog.cache_utils import cache_for
 from posthog.exceptions_capture import capture_exception
 from posthog.helpers.encrypted_fields import EncryptedJSONField
+from posthog.models.github_integration_base import GITHUB_REPOSITORY_CACHE_TTL_SECONDS, GitHubIntegrationBase
 from posthog.models.instance_setting import get_instance_settings
 from posthog.models.oauth import OAuthAccessToken, OAuthApplication, OAuthRefreshToken
 from posthog.models.user import User
@@ -79,33 +77,7 @@ oauth_refresh_counter = Counter(
     "integration_oauth_refresh", "Number of times an oauth refresh has been attempted", labelnames=["kind", "result"]
 )
 
-github_api_request_counter = Counter(
-    "github_integration_api_requests",
-    "Number of GitHub API requests made through a GitHub integration.",
-    labelnames=["integration_id", "method", "endpoint", "status_code"],
-)
-
 GITHUB_API_VERSION = "2022-11-28"
-github_api_rate_limit_remaining_gauge = Gauge(
-    "github_integration_api_rate_limit_remaining",
-    "Most recently observed GitHub API rate limit remaining count by integration and resource.",
-    labelnames=["integration_id", "resource"],
-)
-github_api_rate_limit_limit_gauge = Gauge(
-    "github_integration_api_rate_limit_limit",
-    "Most recently observed GitHub API rate limit limit by integration and resource.",
-    labelnames=["integration_id", "resource"],
-)
-github_api_rate_limit_reset_timestamp_gauge = Gauge(
-    "github_integration_api_rate_limit_reset_timestamp_seconds",
-    "Most recently observed GitHub API rate limit reset timestamp by integration and resource.",
-    labelnames=["integration_id", "resource"],
-)
-github_cache_access_counter = Counter(
-    "github_integration_cache_accesses",
-    "Number of GitHub integration cache accesses by cache type, repository, and result.",
-    labelnames=["integration_id", "cache", "repository", "result"],
-)
 
 PRIVATE_CHANNEL_WITHOUT_ACCESS = "PRIVATE_CHANNEL_WITHOUT_ACCESS"
 
@@ -2068,21 +2040,11 @@ class JiraIntegration:
         return {"key": issue.get("key", ""), "id": issue.get("id", "")}
 
 
-@dataclass(frozen=True)
-class GitHubCommitAuthor:
-    login: str
-    name: str | None
-    commit_url: str
-
-
 # Default branches change rarely; a multi-hour TTL is plenty to avoid hitting
 # GitHub on every paginated branch request while keeping the window in which a
 # renamed default branch stays stale tolerably short.
 GITHUB_DEFAULT_BRANCH_CACHE_TTL_SECONDS = 60 * 60 * 6
-GITHUB_REPOSITORY_CACHE_TTL_SECONDS = 60 * 60
 GITHUB_REPOSITORY_REFRESH_COOLDOWN_SECONDS = 30
-GITHUB_BRANCH_CACHE_TTL_SECONDS = 60 * 10
-GITHUB_BRANCH_CACHE_TIMEOUT_SECONDS = 60 * 60 * 24
 
 
 @dataclass(frozen=True)
@@ -2151,49 +2113,9 @@ def raise_if_github_rate_limited(response: requests.Response) -> None:
     )
 
 
-class GitHubIntegration:
+class GitHubIntegration(GitHubIntegrationBase):
     integration: Integration
 
-    @classmethod
-    def client_request(cls, endpoint: str, method: str = "GET") -> requests.Response:
-        github_app_client_id = settings.GITHUB_APP_CLIENT_ID
-        github_app_private_key = settings.GITHUB_APP_PRIVATE_KEY
-
-        if not github_app_client_id:
-            raise ValidationError("GITHUB_APP_CLIENT_ID is not configured")
-
-        if not github_app_private_key:
-            raise ValidationError("GITHUB_APP_PRIVATE_KEY is not configured")
-
-        github_app_private_key = github_app_private_key.replace("\\n", "\n").strip()
-
-        try:
-            jwt_token = jwt.encode(
-                {
-                    "iat": int(time.time()) - 300,  # 5 minutes in the past
-                    "exp": int(time.time()) + 300,  # 5 minutes in the future
-                    "iss": github_app_client_id,
-                },
-                github_app_private_key,
-                algorithm="RS256",
-            )
-        except Exception:
-            logger.error("Failed to encode JWT token", exc_info=True)
-            raise ValidationError(
-                "Failed to create GitHub App JWT token. Please check your GITHUB_APP_PRIVATE_KEY format."
-            )
-
-        return requests.request(
-            method,
-            f"https://api.github.com/app/{endpoint}",
-            headers={
-                "Accept": "application/vnd.github+json",
-                "Authorization": f"Bearer {jwt_token}",
-                "X-GitHub-Api-Version": GITHUB_API_VERSION,
-            },
-        )
-
-    @classmethod
     def integration_from_installation_id(
         cls, installation_id: str, team_id: int, created_by: User | None = None
     ) -> Integration:
@@ -2318,93 +2240,6 @@ class GitHubIntegration:
             raise Exception("GitHubIntegration init called with Integration with wrong 'kind'")
         self.integration = integration
 
-    @staticmethod
-    def _rate_limit_header(headers: Mapping[str, str] | None, name: str) -> float | None:
-        if headers is None:
-            return None
-        value = headers.get(name)
-        if value is None:
-            return None
-        try:
-            return float(value)
-        except (TypeError, ValueError):
-            return None
-
-    def _record_github_api_response(self, response: requests.Response, method: str, endpoint: str) -> None:
-        integration_id = str(self.integration.id)
-        status_code = str(response.status_code)
-        github_api_request_counter.labels(integration_id, method, endpoint, status_code).inc()
-
-        headers = response.headers if isinstance(response.headers, Mapping) else None
-        resource = headers.get("X-RateLimit-Resource", "unknown") if headers is not None else "unknown"
-        remaining = self._rate_limit_header(headers, "X-RateLimit-Remaining")
-        limit = self._rate_limit_header(headers, "X-RateLimit-Limit")
-        reset_at = self._rate_limit_header(headers, "X-RateLimit-Reset")
-
-        if remaining is not None:
-            github_api_rate_limit_remaining_gauge.labels(integration_id, resource).set(remaining)
-        if limit is not None:
-            github_api_rate_limit_limit_gauge.labels(integration_id, resource).set(limit)
-        if reset_at is not None:
-            github_api_rate_limit_reset_timestamp_gauge.labels(integration_id, resource).set(reset_at)
-
-    def _record_github_api_exception(self, method: str, endpoint: str) -> None:
-        github_api_request_counter.labels(str(self.integration.id), method, endpoint, "exception").inc()
-
-    def _record_github_cache_access(
-        self, cache_type: Literal["repositories", "branches"], result: Literal["hit", "miss"], repository: str
-    ) -> None:
-        github_cache_access_counter.labels(str(self.integration.id), cache_type, repository.casefold(), result).inc()
-
-    def _github_api_get(
-        self,
-        url: str,
-        *,
-        endpoint: str,
-        headers: dict[str, str],
-        params: dict[str, str | int] | None = None,
-        timeout: int | None = None,
-    ) -> requests.Response:
-        try:
-            response = requests.get(url, headers=headers, params=params, timeout=timeout)
-        except requests.RequestException:
-            self._record_github_api_exception("GET", endpoint)
-            raise
-        self._record_github_api_response(response, "GET", endpoint)
-        return response
-
-    def _github_api_post(
-        self,
-        url: str,
-        *,
-        endpoint: str,
-        headers: dict[str, str],
-        json_body: Mapping[str, object] | None = None,
-    ) -> requests.Response:
-        try:
-            response = requests.post(url, json=json_body, headers=headers)
-        except requests.RequestException:
-            self._record_github_api_exception("POST", endpoint)
-            raise
-        self._record_github_api_response(response, "POST", endpoint)
-        return response
-
-    def _github_api_put(
-        self,
-        url: str,
-        *,
-        endpoint: str,
-        headers: dict[str, str],
-        json_body: Mapping[str, object],
-    ) -> requests.Response:
-        try:
-            response = requests.put(url, json=json_body, headers=headers)
-        except requests.RequestException:
-            self._record_github_api_exception("PUT", endpoint)
-            raise
-        self._record_github_api_response(response, "PUT", endpoint)
-        return response
-
     def access_token_expired(self, time_threshold: timedelta | None = None) -> bool:
         expires_in = self.integration.config.get("expires_in")
         refreshed_at = self.integration.config.get("refreshed_at")
@@ -2447,7 +2282,6 @@ class GitHubIntegration:
             reload_integrations_on_workers(self.integration.team_id, [self.integration.id])
             oauth_refresh_counter.labels(self.integration.kind, "success").inc()
             self.integration.save()
-
     def get_access_token(self) -> str:
         """Return a valid installation access token, refreshing it if expired."""
         if self.access_token_expired():
@@ -2456,268 +2290,6 @@ class GitHubIntegration:
         if not token:
             raise GitHubIntegrationError("Access token unavailable after refresh")
         return token
-
-    def organization(self) -> str:
-        return dot_get(self.integration.config, "account.name")
-
-    def _installation_authenticated_get(
-        self, url: str, *, endpoint: str, timeout: int = 10
-    ) -> requests.Response | None:
-        """GET with installation token; refreshes on expiry or 401."""
-        try:
-            if self.access_token_expired():
-                self.refresh_access_token()
-        except Exception:
-            logger.warning("GitHubIntegration: token refresh pre-check failed", exc_info=True)
-
-        def fetch() -> requests.Response:
-            access_token = self.integration.sensitive_config.get("access_token")
-            return self._github_api_get(
-                url,
-                endpoint=endpoint,
-                headers={
-                    "Accept": "application/vnd.github+json",
-                    "Authorization": f"Bearer {access_token}",
-                    "X-GitHub-Api-Version": GITHUB_API_VERSION,
-                },
-                timeout=timeout,
-            )
-
-        try:
-            response = fetch()
-            if response.status_code == 401:
-                try:
-                    self.refresh_access_token()
-                except Exception:
-                    logger.warning("GitHubIntegration: token refresh after 401 failed", exc_info=True)
-                    return None
-                response = fetch()
-            return response
-        except Exception:
-            logger.warning("GitHubIntegration: installation GET failed", url=url, exc_info=True)
-            return None
-
-    def installation_can_access_repository(self, repository: str) -> bool:
-        """Whether this installation token can access the repo (``GET /repos/{owner}/{repo}`` returns 200)."""
-        response = self._installation_authenticated_get(
-            f"https://api.github.com/repos/{repository}", endpoint="/repos/{owner}/{repo}"
-        )
-        if response is None:
-            return False
-        return response.status_code == 200
-
-    def get_commit_author_info(self, repository: str, sha: str) -> GitHubCommitAuthor | None:
-        """Resolve a commit SHA to author metadata via the GitHub API."""
-        response = self._installation_authenticated_get(
-            f"https://api.github.com/repos/{repository}/commits/{sha}", endpoint="/repos/{owner}/{repo}/commits/{sha}"
-        )
-        if response is None:
-            return None
-        if response.status_code != 200:
-            logger.info(
-                "GitHub API non-200 for commit lookup",
-                status_code=response.status_code,
-                sha_prefix=sha[:8],
-                repository=repository,
-            )
-            return None
-        try:
-            data = response.json()
-        except Exception:
-            logger.warning(
-                "GitHubIntegration: failed to parse commit JSON",
-                repository=repository,
-                sha_prefix=sha[:8],
-                exc_info=True,
-            )
-            return None
-        author = data.get("author")
-        if not author or not author.get("login"):
-            return None
-        git_author = data.get("commit", {}).get("author", {})
-        name = git_author.get("name") or author.get("login")
-        commit_url = data.get("html_url", f"https://github.com/{repository}/commit/{sha}")
-        return GitHubCommitAuthor(login=author["login"], name=name, commit_url=commit_url)
-
-    def list_repositories(self, *, limit: int = 100, offset: int = 0) -> tuple[list[dict], bool]:
-        """List installation repositories via the GitHub API.
-
-        Fetches only the GitHub pages needed to satisfy the requested
-        ``[offset, offset+limit)`` window. Returns a tuple of
-        ``(repositories, has_more)`` where *has_more* indicates whether
-        additional repositories exist beyond the returned window.
-        """
-        GITHUB_PER_PAGE = 100
-
-        try:
-            if self.access_token_expired():
-                self.refresh_access_token()
-        except Exception:
-            logger.warning("GitHubIntegration: token refresh pre-check failed", exc_info=True)
-
-        def fetch(page: int) -> requests.Response:
-            access_token = self.integration.sensitive_config.get("access_token")
-            return self._github_api_get(
-                f"https://api.github.com/installation/repositories?page={page}&per_page={GITHUB_PER_PAGE}",
-                endpoint="/installation/repositories",
-                headers={
-                    "Accept": "application/vnd.github+json",
-                    "Authorization": f"Bearer {access_token}",
-                    "X-GitHub-Api-Version": GITHUB_API_VERSION,
-                },
-            )
-
-        def extract_repos(body: dict) -> list[dict]:
-            repositories = body.get("repositories")
-            if not isinstance(repositories, list):
-                logger.warning(
-                    "GitHubIntegration: list_repositories invalid payload",
-                    integration_id=self.integration.id,
-                    payload_keys=sorted(body.keys()),
-                )
-                raise GitHubIntegrationError("GitHubIntegration: list_repositories invalid payload")
-            return [
-                {
-                    "id": repo["id"],
-                    "name": repo["name"],
-                    "full_name": repo["full_name"],
-                }
-                for repo in repositories
-                if isinstance(repo, dict)
-                and isinstance(repo.get("id"), int)
-                and isinstance(repo.get("name"), str)
-                and isinstance(repo.get("full_name"), str)
-            ]
-
-        def raise_repository_error(message: str, *, status_code: int | None = None, exc_info: bool = False) -> None:
-            logger.warning(
-                message,
-                integration_id=self.integration.id,
-                status_code=status_code,
-                exc_info=exc_info,
-            )
-            raise GitHubIntegrationError(message)
-
-        # Work out which GitHub pages cover the requested window.
-        first_page = offset // GITHUB_PER_PAGE + 1
-        skip = offset % GITHUB_PER_PAGE
-        needed = skip + limit
-
-        # Fetch the first required page with 401-retry and transient-error retry.
-        transient_status_codes = {502, 503, 504}
-        current_page = first_page
-
-        for attempt in range(2):
-            try:
-                response = fetch(current_page)
-            except requests.RequestException:
-                raise_repository_error("GitHubIntegration: list_repositories network error", exc_info=True)
-
-            if response.status_code == 401:
-                try:
-                    self.refresh_access_token()
-                except Exception:
-                    raise_repository_error("GitHubIntegration: token refresh after 401 failed", exc_info=True)
-                try:
-                    response = fetch(current_page)
-                except requests.RequestException:
-                    raise_repository_error("GitHubIntegration: list_repositories network error on retry", exc_info=True)
-
-            try:
-                body = response.json()
-            except Exception:
-                if response.status_code in transient_status_codes and attempt == 0:
-                    logger.info(
-                        "GitHubIntegration: list_repositories retrying transient non-JSON response",
-                        status_code=response.status_code,
-                    )
-                    continue
-                logger.warning(
-                    "GitHubIntegration: list_repositories non-JSON response",
-                    integration_id=self.integration.id,
-                    status_code=response.status_code,
-                )
-                raise GitHubIntegrationError("GitHubIntegration: list_repositories non-JSON response")
-
-            if response.status_code == 200 and isinstance(body, dict):
-                page_repos = extract_repos(body)
-                all_fetched = page_repos
-                has_next_page = len(page_repos) == GITHUB_PER_PAGE
-                break
-
-            if response.status_code in transient_status_codes and attempt == 0:
-                logger.info(
-                    "GitHubIntegration: list_repositories retrying transient error",
-                    status_code=response.status_code,
-                    error=body if isinstance(body, dict) else None,
-                )
-                continue
-
-            logger.warning(
-                "GitHubIntegration: failed to list repositories",
-                integration_id=self.integration.id,
-                status_code=response.status_code,
-                error=body if isinstance(body, dict) else None,
-            )
-            raise GitHubIntegrationError("GitHubIntegration: failed to list repositories")
-        else:
-            raise GitHubIntegrationError("GitHubIntegration: failed to list repositories after retries")
-
-        # Fetch subsequent pages until we have enough items.
-        while len(all_fetched) < needed and has_next_page:
-            current_page += 1
-            try:
-                response = fetch(current_page)
-            except requests.RequestException:
-                logger.warning(
-                    "GitHubIntegration: list_repositories network error on page",
-                    integration_id=self.integration.id,
-                    page=current_page,
-                    exc_info=True,
-                )
-                raise GitHubIntegrationError("GitHubIntegration: list_repositories network error on page")
-            try:
-                body = response.json()
-            except Exception:
-                logger.warning(
-                    "GitHubIntegration: list_repositories non-JSON response on page",
-                    integration_id=self.integration.id,
-                    page=current_page,
-                    status_code=response.status_code,
-                )
-                raise GitHubIntegrationError("GitHubIntegration: list_repositories non-JSON response on page")
-            if response.status_code != 200 or not isinstance(body, dict):
-                logger.warning(
-                    "GitHubIntegration: failed to list repositories on page",
-                    integration_id=self.integration.id,
-                    page=current_page,
-                    status_code=response.status_code,
-                    error=body if isinstance(body, dict) else None,
-                )
-                raise GitHubIntegrationError("GitHubIntegration: failed to list repositories on page")
-            page_repos = extract_repos(body)
-            all_fetched.extend(page_repos)
-            has_next_page = len(page_repos) == GITHUB_PER_PAGE
-
-        result = all_fetched[skip : skip + limit]
-        has_more = has_next_page or (skip + limit < len(all_fetched))
-
-        return result, has_more
-
-    def list_all_repositories(self) -> list[dict]:
-        """Fetch all accessible repositories, paginating through GitHub's API."""
-        all_repositories: list[dict] = []
-        offset = 0
-        page_size = 100
-
-        while True:
-            repositories, has_more = self.list_repositories(limit=page_size, offset=offset)
-            all_repositories.extend(repositories)
-
-            if not has_more or not repositories:
-                return all_repositories
-
-            offset += len(repositories)
 
     def _get_repository_cache(self) -> list[dict] | None:
         cached = self.integration.repository_cache
@@ -2847,289 +2419,6 @@ class GitHubIntegration:
     def list_all_cached_repositories_async(self, max_repos: int | None = None) -> list[dict]:
         return self.list_all_cached_repositories(max_repos=max_repos)
 
-    def _get_branch_cache_key(self, repo: str) -> str:
-        return f"github_integration:branches:{self.integration.id}:{repo.lower()}"
-
-    def _get_branch_cache(self, repo: str) -> dict[str, Any] | None:
-        cached = cache.get(self._get_branch_cache_key(repo))
-        if not isinstance(cached, dict):
-            return None
-
-        branches = cached.get("branches")
-        default_branch = cached.get("default_branch")
-        updated_at = cached.get("updated_at")
-        if not isinstance(branches, list) or not all(isinstance(branch, str) for branch in branches):
-            return None
-        if default_branch is not None and not isinstance(default_branch, str):
-            return None
-        if not isinstance(updated_at, (int, float)):
-            return None
-
-        return {
-            "branches": branches,
-            "default_branch": default_branch,
-            "updated_at": updated_at,
-        }
-
-    def branch_cache_is_stale(self, repo: str) -> bool:
-        cached = self._get_branch_cache(repo)
-        if cached is None:
-            return True
-
-        return time.time() - float(cached["updated_at"]) >= GITHUB_BRANCH_CACHE_TTL_SECONDS
-
-    def list_all_branches(self, repo: str) -> list[str]:
-        """Fetch all branches for a repository, paginating through GitHub's API."""
-        all_branches: list[str] = []
-        offset = 0
-        page_size = 100
-
-        while True:
-            branches, has_more = self.list_branches(repo, limit=page_size, offset=offset)
-            all_branches.extend(branches)
-
-            if not has_more or not branches:
-                return all_branches
-
-            offset += len(branches)
-
-    def sync_branch_cache(self, repo: str) -> tuple[list[str], str | None]:
-        branches = self.list_all_branches(repo)
-        cached = self._get_branch_cache(repo)
-        cached_default_branch = None if cached is None else cast(str | None, cached["default_branch"])
-
-        default_branch: str | None
-        try:
-            default_branch = self.get_default_branch(repo)
-        except Exception:
-            logger.warning(
-                "GitHubIntegration: failed to refresh default branch",
-                integration_id=self.integration.id,
-                repo=repo,
-                exc_info=True,
-            )
-            default_branch = cached_default_branch if cached_default_branch in branches else None
-
-        if default_branch and default_branch in branches:
-            branches = [branch for branch in branches if branch != default_branch]
-            branches.insert(0, default_branch)
-
-        cache.set(
-            self._get_branch_cache_key(repo),
-            {
-                "branches": branches,
-                "default_branch": default_branch,
-                "updated_at": time.time(),
-            },
-            timeout=GITHUB_BRANCH_CACHE_TIMEOUT_SECONDS,
-        )
-
-        return branches, default_branch
-
-    def list_cached_branches(
-        self, repo: str, *, search: str = "", limit: int = 100, offset: int = 0
-    ) -> tuple[list[str], str | None, bool]:
-        cached = self._get_branch_cache(repo)
-        should_refresh = cached is None or self.branch_cache_is_stale(repo)
-        self._record_github_cache_access("branches", "miss" if should_refresh else "hit", repo)
-
-        if should_refresh:
-            try:
-                branches, default_branch = self.sync_branch_cache(repo)
-                cached = {
-                    "branches": branches,
-                    "default_branch": default_branch,
-                }
-            except Exception:
-                logger.warning(
-                    "GitHubIntegration: failed to refresh branch cache",
-                    integration_id=self.integration.id,
-                    repo=repo,
-                    exc_info=True,
-                )
-                if cached is None:
-                    raise
-
-        assert cached is not None
-        branches = cast(list[str], cached["branches"])
-        default_branch = cast(str | None, cached["default_branch"])
-
-        normalized_search = search.strip().casefold()
-        filtered_branches = (
-            [branch for branch in branches if normalized_search in branch.casefold()] if normalized_search else branches
-        )
-
-        result = filtered_branches[offset : offset + limit]
-        has_more = offset + limit < len(filtered_branches)
-        return result, default_branch, has_more
-
-    def get_top_starred_repository(self) -> str | None:
-        """Get the repository with the most stars from the GitHub integration.
-
-        Returns the full repository name in format 'org/repo', or None if no repos available.
-        """
-        try:
-            if self.access_token_expired():
-                self.refresh_access_token()
-        except Exception:
-            logger.warning("GitHubIntegration: token refresh pre-check failed", exc_info=True)
-
-        def fetch(page: int = 1) -> requests.Response:
-            access_token = self.integration.sensitive_config.get("access_token")
-            return self._github_api_get(
-                f"https://api.github.com/installation/repositories?page={page}&per_page=100",
-                endpoint="/installation/repositories",
-                headers={
-                    "Accept": "application/vnd.github+json",
-                    "Authorization": f"Bearer {access_token}",
-                    "X-GitHub-Api-Version": GITHUB_API_VERSION,
-                },
-            )
-
-        response = fetch()
-
-        if response.status_code == 401:
-            try:
-                self.refresh_access_token()
-            except Exception:
-                logger.warning("GitHubIntegration: token refresh after 401 failed", exc_info=True)
-            else:
-                response = fetch()
-
-        try:
-            body = response.json()
-        except Exception:
-            logger.warning(
-                "GitHubIntegration: get_top_starred_repository non-JSON response",
-                status_code=response.status_code,
-            )
-            return None
-
-        repositories = body.get("repositories")
-        if response.status_code != 200 or not isinstance(repositories, list) or not repositories:
-            return None
-
-        top_repo = max(repositories, key=lambda r: r.get("stargazers_count", 0) if isinstance(r, dict) else 0)
-        if not isinstance(top_repo, dict):
-            return None
-
-        full_name = top_repo.get("full_name")
-        if isinstance(full_name, str):
-            return full_name.lower()
-
-        return None
-
-    def list_branches(self, repo: str, *, limit: int = 100, offset: int = 0) -> tuple[list[str], bool]:
-        """List branches for a given repository via the GitHub API.
-
-        Fetches only the GitHub pages needed to satisfy the requested
-        ``[offset, offset+limit)`` window. Returns a tuple of
-        ``(branch_names, has_more)`` where *has_more* indicates whether
-        additional branches exist beyond the returned window.
-        """
-        GITHUB_PER_PAGE = 100
-
-        try:
-            if self.access_token_expired():
-                self.refresh_access_token()
-        except Exception:
-            logger.warning("GitHubIntegration: token refresh pre-check failed", exc_info=True)
-
-        def fetch(page: int) -> requests.Response:
-            access_token = self.integration.sensitive_config.get("access_token")
-            return self._github_api_get(
-                f"https://api.github.com/repos/{repo}/branches?per_page={GITHUB_PER_PAGE}&page={page}",
-                endpoint="/repos/{owner}/{repo}/branches",
-                headers={
-                    "Accept": "application/vnd.github+json",
-                    "Authorization": f"Bearer {access_token}",
-                    "X-GitHub-Api-Version": GITHUB_API_VERSION,
-                },
-                timeout=10,
-            )
-
-        def extract_names(data: list) -> list[str]:
-            return [
-                branch["name"] for branch in data if isinstance(branch, dict) and isinstance(branch.get("name"), str)
-            ]
-
-        # Work out which GitHub pages cover the requested window.
-        first_page = offset // GITHUB_PER_PAGE + 1
-        skip = offset % GITHUB_PER_PAGE
-        needed = skip + limit
-
-        # Fetch the first required page (with 401-retry logic).
-        current_page = first_page
-        try:
-            response = fetch(current_page)
-        except requests.RequestException:
-            logger.warning("GitHubIntegration: list_branches network error", repo=repo, exc_info=True)
-            return [], False
-
-        if response.status_code == 401:
-            try:
-                self.refresh_access_token()
-            except Exception:
-                logger.warning("GitHubIntegration: token refresh after 401 failed", exc_info=True)
-                return [], False
-            try:
-                response = fetch(current_page)
-            except requests.RequestException:
-                logger.warning("GitHubIntegration: list_branches network error on retry", repo=repo, exc_info=True)
-                return [], False
-
-        if response.status_code != 200:
-            logger.warning(
-                "GitHubIntegration: failed to list branches",
-                status_code=response.status_code,
-                repo=repo,
-            )
-            return [], False
-
-        try:
-            body = response.json()
-        except Exception:
-            logger.warning(
-                "GitHubIntegration: list_branches non-JSON response",
-                status_code=response.status_code,
-            )
-            return [], False
-
-        if not isinstance(body, list):
-            return [], False
-
-        all_fetched = extract_names(body)
-        has_next_page = 'rel="next"' in response.headers.get("Link", "")
-
-        # Fetch subsequent pages until we have enough items.
-        while len(all_fetched) < needed and has_next_page:
-            current_page += 1
-            try:
-                response = fetch(current_page)
-            except requests.RequestException:
-                break
-            if response.status_code != 200:
-                logger.warning(
-                    "GitHubIntegration.list_branches pagination stopped",
-                    status_code=response.status_code,
-                    page=current_page,
-                    repo=repo,
-                )
-                break
-            try:
-                body = response.json()
-            except Exception:
-                break
-            if not isinstance(body, list):
-                break
-            all_fetched.extend(extract_names(body))
-            has_next_page = 'rel="next"' in response.headers.get("Link", "")
-
-        result = all_fetched[skip : skip + limit]
-        has_more = has_next_page or (skip + limit < len(all_fetched))
-
-        return result, has_more
-
     def create_issue(self, config: dict[str, str]):
         title: str = config.pop("title")
         body: str = config.pop("body")
@@ -3152,38 +2441,6 @@ class GitHubIntegration:
         issue = response.json()
 
         return {"number": issue["number"], "repository": repository}
-
-    def get_default_branch(self, repository: str) -> str:
-        """Get the default branch for a repository."""
-        repo_path = repository if "/" in repository else f"{self.organization()}/{repository}"
-        cache_key = f"github_integration:default_branch:{self.integration.id}:{repo_path}"
-
-        cached = cache.get(cache_key)
-        if isinstance(cached, str):
-            return cached
-
-        access_token = self.integration.sensitive_config.get("access_token")
-        if not access_token:
-            raise ValueError("GitHub access token not configured")
-
-        response = self._github_api_get(
-            f"https://api.github.com/repos/{repo_path}",
-            endpoint="/repos/{owner}/{repo}",
-            headers={
-                "Accept": "application/vnd.github+json",
-                "Authorization": f"Bearer {access_token}",
-                "X-GitHub-Api-Version": GITHUB_API_VERSION,
-            },
-            timeout=10,
-        )
-
-        if response.status_code == 200:
-            repo_data = response.json()
-            default_branch = repo_data.get("default_branch", "main")
-            cache.set(cache_key, default_branch, timeout=60 * 60 * 24)
-            return default_branch
-        else:
-            raise Exception(f"Failed to get default branch: HTTP {response.status_code}")
 
     def create_branch(self, repository: str, branch_name: str, base_branch: str | None = None) -> dict[str, Any]:
         """Create a new branch from a base branch."""
